fix errors and add new
